+++ lora-e5-bsp/src/lib.rs
@@ -12,8 +12,12 @@ use hal::{
     gpio::{self, pins, Output, OutputArgs, PinState},
 };
 
+#[cfg(feature = "defmt")]
+use dfmt as defmt;
+
 /// RF switch
 #[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct RfSwitch {
     a4: Output<pins::A4>,
     a5: Output<pins::A5>,