+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `info::CORE` to get the CPU core at compile time.
 - Added `info::core()` to get the CPU core at runtime.
 - Added a `flash` module with erase and program functionality.
+- Added `defmt::Format` for all types declared in the BSPs.
 
 ## [0.2.1] - 2021-11-20
 ### Fixed